Import initial version of spotcat
This is a simple tool using the despotify library to read compressed audio
data for a given list of album or track URIs from Spotify. It is possible to
use this for streaming playback with

  spotcat USERNAME PASSWORD URI [URI]... | ogg123 -q -

Tracks will always be returned in Ogg Vorbis format at the highest supported
bit-rate, as neither DECODERS nor MP3_SUPPORT are defined when building the
despotify library.